requirements.txt
streamlit
yfinance
pandas
plotly
google-generativeai
pytz
requests